requirements.txt
python-telegram-bot==20.7
feedparser==6.0.10
apscheduler==3.10.4
deep-translator==1.11.4
